Refactor application for better code organization and usability
The main changes include improving tool bar component to accept props for the right and left parts of the bar, removing Auth component - the logic for Login and Register are now handled separately in their own pages. New endpoint added for Products to allow for product management. Storage and Shelves files are revised to improve the layout.  The changes are aimed at enhancing code readability, scalability and overall user experience.