Implement features and adjustments for Products feature
Included change:
- Introduced new product schemas
- Imported and utilized ProductCard instead of JSON.stringify()
- Removed unused `Body1Typography` component
- Created new ProductInfoModal component
- Added and implemented deleteProduct, updateProduct and createProduct API end points.
- Min value of weight and size attributes in CreateShelfDtoSchema updated.
- Adjusted styling and added hover effect for ProductCard component
- Updated schemas for validation
- Included attributes and values in the product creation modal
- Added some attributes in getTextFieldProps to handle new field inputs

The changes were necessary to ensure the correct handling of products in the application including its creation, editing, deletion, and also its visual representation and associated data validation. The introduced schemas ensure correct validation of user input giving a better user experience and prevention of incorrect data. The ProductCard offers a better representation of the product. The createProduct, updateProduct, and deleteProduct end points allow manipulation of product data in the backend.