Introduce NoDataMessage component to handle empty data sets
Added a new component called NoDataMessage which is used to display a message when a data set is empty in various sections including products, storages, shelves and product tables. This was implemented to provide a more consistent and reusable way of handling empty data across the application, enhancing the user experience and maintaining a clean codebase. The gap in the AppHeader was also reduced for better styling.